\name{xmpp_auth_sasl_digest_md5}
\alias{xmpp_auth_sasl_digest_md5}
\title{authenticate to XMPP server with SASL DIGEST-MD5}
\usage{
  xmpp_auth_sasl_digest_md5(username, password, host, port)
}
\arguments{
  \item{username}{}

  \item{password}{}

  \item{host}{}

  \item{port}{}
}
\value{
  socket connection
}
\description{
  authenticate to XMPP server with SASL DIGEST-MD5
}

