\docType{data}
\name{XMPP_LEVEL_DEBUG}
\alias{XMPP_LEVEL_DEBUG}
\alias{XMPP_LEVEL_ERROR}
\alias{XMPP_LEVEL_INFO}
\alias{XMPP_LEVEL_WARN}
\title{Level for displaying logs}
\format{integer}
\description{
  Level for displaying logs
}
\seealso{
  \code{\link{sendXMPPMessage}}
  \code{\link{facebookMessage}}
}
\keyword{datasets}

